+++ tests/test_util_fromstr.py
@@ -33,10 +33,10 @@ def test_to_bool() -> None:
         ('10 kbb',       ValueError('Filesize string does not match expected pattern')),
     ],
 )
-def test_filesize(s: str, expected: int | Exception) -> None:
+def test_to_filesize(s: str, expected: int | Exception) -> None:
     for i in (s, s.upper()):
         if isinstance(expected, Exception):
             with pytest.raises(expected.__class__, match=expected.args[0]):
-                FromStr.filesize(i)
+                FromStr.to_filesize(i)
         else:
-            assert FromStr.filesize(i) == expected
+            assert FromStr.to_filesize(i) == expected